backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-socketio>=5.11.0
python-dotenv>=1.0.1
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.22
langgraph>=0.0.24
edge-tts>=6.1.9
pydantic>=2.6.1
openai-whisper
